requirements.txt
Django~=4.1.7
django-qrcode~=0.3
django-phonenumber-field~=7.1.0
qrcode>=7.4.2
user-agents>=2.2.0
Pillow>=9.4.0
django-dirtyfields>=1.9.2